refactor: Move initialization related exceptions to data.exception.initialization
